// apps/frontend/src/app/replay/models/error-messages.model.ts
export interface ErrorMessages {
  QueryError: string;
  ParamsError: string;
  401: string;
  UnitIdError: string;
  TestPersonError: string;
  PlayerError: string;
  ResponsesError: string;
  notInList: string;
  notCurrent: string;
  unknown: string;
  tokenExpired: string;
  tokenInvalid: string;
}